requirements.txt
requests>=2.28.2
pyjwt>=2.6.0
numpy>=1.24.2
pandas>=1.5.3
python-dateutil>=2.8.2
pytz>=2022.7.1
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.20.0
plotly>=5.13.0